+++ src/geometry.hpp
@@ -23,7 +23,7 @@ namespace ch {
     point mean_point(const polyline& poly);
     polyline transform(const polyline& poly, const matrix& mat);
     point transform(const point& pt, const matrix& mat);
-    void paint_polyline(cv::Mat& mat, const polyline& p, int thickness, int color, point offset = { 0,0 });
+    void paint_polyline(cv::Mat& mat, const polyline& p, double thickness, int color, point offset = { 0,0 });
 
     double euclidean_distance(const point& pt1, const point& pt2);
